lib: nrf_modem_lib: Add Kconfig for application-specific fault handling
Remove the weak implementation of the fault handler and require the
application to set CONFIG_NRF_MODEM_LIB_ON_FAULT_RESET_MODEM if an
application-specific fault handling is required. This ensures that the
NRF_MODEM_LIB_ON_FAULT Kconfig choice is valid, regardless of whether
the application has a custom fault handler or not.

Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>