manifest: nrf_modem: v2.1.0
Updates for Modem library v2.1.0.

See the changelog in nrfxlib for a detailed list of changes
in this Modem library release.

This patch updates the nrf_modem_lib_trace component to use
the newly introduced nrf_modem_trace_get() to retrieve traces.
Traces are printed by a thread using the trace_backend interface.

The trace thread is automatically started when traces are enabled,
and it uses with libmodem's initialization callbacks to
automatically resume tracing when the modem is re-initialized,
due to a fault or other.

The KConfig option CONFIG_NRF_MODEM_LIB_TRACE_ENABLED was
renamed to CONFIG_NRF_MODEM_LIB_TRACE.

The following options were removed:
- CONFIG_NRF_MODEM_LIB_TRACE_THREAD_PROCESSING
- CONFIG_NRF_MODEM_LIB_TRACE_HEAP_SIZE
- CONFIG_NRF_MODEM_LIB_TRACE_HEAP_SIZE_OVERRIDE
- CONFIG_NRF_MODEM_LIB_TRACE_HEAP_DUMP_PERIODIC
- CONFIG_NRF_MODEM_LIB_TRACE_HEAP_DUMP_PERIOD_MS

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>
Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>
Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>
Signed-off-by: divya pillai <divya.pillai@nordicsemi.no>