manifest: nrf_modem: v2.1.0
Updates for Modem library v2.1.0.

See the changelog in nrfxlib for a detailed list of changes
in this Modem library release.

This patch updates the nrf_modem_lib_trace component
to use the newly introduced nrf_modem_trace_get() function
to retrieve traces. Traces are printed by a thread
that is automatically started when traces are enabled.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>