[nrf noup] samples: coprocessor: Added timing compatibility for Haltium
To run OpenThread / Matter samples to work with FPGA Haltium
there was a need to increase AIFS time without editing the
defines within the nrfxlib repo.

In this commit, two Kconfigs have been added to the example:
the NRF_802154_AIFS_TIME to set AIFS time, and the
HALTIUM_TIMING_COMPATIBILITY to enable Haltium timing compatibility.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>