NRFX-2547: Adding Halti targets to the SysCtrl app
To help with starting the introduction of Halti support, it was
decided that there should be an initial version of the Halti targets.
This allows development and testing together with the simulation tools,
and allow the rest of the team to work on Halti support.

To achieve this, the following changes were made:
- Add builable targets for SysCtrl:
  - nrf92sysc_nrf9230
  - nrf92emu_nrf9230
  - nrf92fpga_nrf9230
- Add support for NRF9230 targets to Zephyr's drivers
- Add support for NRF9230 targets to the SysCtrl SDK

Signed-off-by: Saku Rautio (Ext) <saku.rautio@nordicsemi.no>