test-spec: build docs when manifest is updated
Internal documentation is in a separated repository linked with NCS docs
through west.yml manifest. When this documentation is updated docs
should be built.

The only way of detecting changes in external repositories by test-spec
is to detect any change in west.yml manifest. Because of this west.yml
is added to the doc build label. The accepted cosequence is building
documentation for any change in dependent repositories even if such
changes do not modify documentation.

Signed-off-by: Hubert Miś <hubert.mis@nordicsemi.no>