modules: nrf_802154: Fix timestamping for multiprotocol on nRF54H
In multiprotocol applications (802.15.4 + BLE) on nRF54H
the CHEN bit for the PPI_TIMESTAMP channel is cleared by BLE.
This, combined with the fact that until know this CHEN bit was
only enabled at the initialization stage, causes that the
timestamping in the 802.15.4 driver was not triggered correctly
and the timestamps are 0.

This fixes the issue by moving the enabling of the channel to
the start of the 802.15.4 timeslot.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>