ncs: partial upmerge
This commit is an attempt to cherry-pick all changes to NCS made by
the Aegir team since v2.4.0 until the f2a8e3c8f revision of nrf.

This is done to reduce the amount of conflicts that need to be
resolved in new code developed in HCS.

There were 148 commits to NCS in this period and cherry-picking 148
commits was considered to be too slow.

So instead, each file touched by the team was checked out to the new
revision. This was done with the following commands:

git log --author="Sebastian B" --author="Georgios" --author=Vidar
--author="Joakim A" --author="Frank" --author="Markus Sw" --format=""
--name-only latest_hcs_upmerge_sync..new_hcs_cherry_pick_point | sort
| uniq

git checkout new_hcs_cherry_pick_point
--pathspec-from-file=pathspec.txt

This strategy is faster than 148 cherry-picks, but has the unfortunate
side-effect of pulling in changes that our team did not make.

When resolving conflicts from this strategy some patches were
accidentally left out and then later cherry-picked with the fromtree
tag.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>