manifest: Update Find My revision
Prepare manifest for Find My repository change to the -next specific
one.

It is necessary to change the Find My revision from the tag to the
commit hash before changing the repo-path itself, as the new repo does
not have this tag created and the CI fails on "Test Spec & Select
downstream" step.

Jira: NCSDK-22502

Signed-off-by: Mateusz Kapala <mateusz.kapala@nordicsemi.no>