[nrf fromtree] debug: etb_trace: use z_get_next_timeout_expiry
Instead of accessing _kernel.idle, only available if CONFIG_PM=y which
is no longer selectable. Note that _kernel.idle is assigned with the
value of z_get_next_timeout_expiry in the idle thread, so the code
should behave the same way as before. Usage of internal Zephyr APIs is
still dangerous, but this can be fixed later.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>