samples: ipc: ipc_service: Add configurable gpio to measure ipc latency
This commit adds configurable GPIOs that would change its state to high
when the ipc_service_send command is processing the data
and other pin that would change its state when the data is being
processed on the ipc receive callback.

The commit also implements 2 additional configurations
that allows checking of the ipc latency between radio and app core.

Signed-off-by: Radoslaw Koppel <radoslaw.koppel@nordicsemi.no>