[nrf noup] nrf_security: Cracen, hard-code SECP521 R1
This explicitely sets the CONFIG_PSA_ACCEL_KEY_PAIR_R521
define for Cracen. This is a hack to allow the Oberon
PSA core to build. Since the R1 SECP521 curve is not
supported by Oberon if we don't set that we will get a
build error. This is intended to be reverted when Cracen
will have support for the new configuration system of
Oberon PSA_NEED_* and PSA_ACCEL_* will be set accordingly.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>