applications: psn: Temporary change to 64Mhz clock
Turned out, after xx was merged it broke spi and i2c.
spi00 won't work until NRFX-3331 is int.
For i2c22 NRFX-4580 is reported.

To omit some of those issues, temporary reduce clock
speed to 64Mhz.

@nika-nordic said softdevice controller claim bluetooth
will not be funcional with 64Mhz, but this is a tradeof
for now.

Signed-off-by: Emil Obalski <Emil.Obalski@nordicsemi.no>