fatal_error: Disable reset on Haltium platforms
Allowing local domains to trigger a reset request when encountering a
fault or fatal operation can lead to a state in which the device can
no longer be accessible from nrfjprog and requires a recovery
procedure. This may be related to the Secure Domain seeing local
domain faults as normal reset requests, and then servicing them in a
rapid loop forever.

Until a more robust solution to the such reset looping can be
determined, prevent all Haltium CPUs from enabling such a reset by
default.

Ref: NCSDK-24373

Signed-off-by: Stephen Stauts <stephen.stauts@nordicsemi.no>