dts: rename from nordic_rng to rng
Define the fake entropy-prng driver to be named rng instead of
nordic_rng.

rng is also the name of the NRF_RNG driver. Re-using this name allows
pre-existing SW that needs an NRF_RNG to get our fake entropy instead.

This will be replaced with PSA entropy when available (NCSDK-13168).

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>