mpsl: flash sync: Do not use previous write times on 54l
The time taken in a previous write is not a predictor of the time
taken for the next write. Writing the same value as is already stored
is much faster than writing a different value. If the first few
writes are fast and the later ones are slow we may get an
overstay assert. The configured event length is only guaranteed
to fit one write block.

Signed-off-by: Herman Berget <herman.berget@nordicsemi.no>