[nrf fromtree] tfm: Add non-secure fault handler callback service
This commit adds a secure service in the platform partition that the
nonsecure application can use to install a callback function to be
called when a fault is triggered in TF-M by the non-secure application.
Since a nonsecure function call from nonsecure cannot give any arguments
the function arguments needs to be provided as non-secure memory area
for the secure service to write to before calling the callback.

The function only calls the handlers for BusFault, SecureFault and
SPU interrupt as these are not banked and can only occur in either
secure or non-secure.
We don't handle hardfaults as this can only happen when the more
specific fault handlers themselves faults.

Co-authored-by: Chris Coleman <chris@memfault.com>
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 544128a2d4687864a7e2477417c1bc356ac872a6)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>