suit: digest cache - compare instead of getting digest
Due to the usage scenario of the digest cache it is both
safer and more convenient to compare a given digest against
the digest stored in cache instead of of getting the digest
from the cache.

This commit replaces suit_plat_digest_cache_get
with suit_plat_digest_cache_compare.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>