suit: removed locking/unlocking of the digest cache
The reason why locking/unlocking of the cache
was used was to limit the number of places where
digests are saved - when the cache is filled
no new elements can be added so the cache was supposed
only to be written to when explicitly allowed by the higher
layer.

However, due to the fact that the cache has a pretty large
amount of elements (it can still be increased if it turns
out it is to small) - 16 - and this does not take a lot
of RAM (as the cache only stores pointers to digests and
component IDs) the usage of this is pretty limitted.

On the other hand controlling the permissions to the cache
from outside of the cache could be potentially dangerous,
so the feature is removed.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>